Refactor FormioReactComponent to remove boilerplates for implementing components
This makes sure all React components will have fixes for initializing with submission, and behave the same. After testing, it seems to work quite well.
The goal is also that it should require as little as possible to create a new component.
